requirements.txt
requests
ipapi
selenium-wire
undetected_chromedriver
blinker==1.7.0
